Split user API
